Mutations module. Moving classes there.
